Added command execution time to the prompt ⏱️🚀
When slower than 5s, the prompt now shows how long the last command took to
run. The duration is displayed as a tutle emoji followed by the time it took
to run. If less than a minute, it'll show seconds elapsed. Otherwise it'll
show minutes and seconds.

We implemented this by adding a preexec hook that records the start
time of the command in milliseconds and a precmd hook that calculates the
duration and sets the prompt time variable accordingly.

This will be helpful for identifying slow commands and flagging them
for possible optimization.